+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 ## Next Release
 
+## 0.0.45
+
+- [sdk] Attempt fix on trust wallet again.
+
+## 0.0.44
+
+- [sdk] Attempt fix on trust wallet again.
+
+## 0.0.41
+
+- [sdk] Attempt fix on trust wallet again.
 - [sdk] Estimate gas on SDK transactions.
 - [router] Serialize tx submission.